Safe Area 対応の修正
