+++ web/app/home/layout.tsx
@@ -11,7 +11,7 @@ export default function Layout({
     <>
       <Header />
       <NavigateByAuthState type="toLoginForUnauthenticated">
-        <div className="h-full pt-12 pb-12">{children}</div>
+        <div className="cm-pb-footer h-full pt-12">{children}</div>
       </NavigateByAuthState>
       <BottomBar activeTab="0_home" />
     </>
